Fix/cleanup timing logic in shakeRun
There were a number of issues here

1. The time was just wrong if called via runAction as it timed how long
the original action took rather than how fast runAction returned which
could be quite a bit before in the days of `kick`. Now `kick` is
removed, `runAction` is unecessary so this also removes the complication
to do with the barrier.

2. Time taken to write the profile was included in the time. (This is
significant in my tests, 0.2s approx on GHC)