Support PackageImports with multi components
The fix is quite hacky. After parsing we post-process the module to
remove any package imports for packages which are in our component.

This is ok, because we already decided ourselves which files these
imports refer to whilst taking into account package imports (see `locateModule`).

If you don't do this processing then GHC will happily decide for itself
again that you really meant to use the version of the package from the
package database rather than the file in the home module. Hopefully
fendor will fix this in GHC this summer.